did some reorganization of files and names
