+++ src/InGame.lua
@@ -9,23 +9,23 @@ local Camera = require "camera"
 -- Signal = require "signal"
 local HC = require "HC"
 
-local Actor = require "Actor"
-local PlayerController = require "PlayerKeymap"
+local Object = require "Object"
+local PlayerController = require "PlayerController"
 --local Gamestate = require "gamestate"
 
 --local table
 local state = {}
 
-local map = require "map"
+local Map = require "Map"
 local cam
 
 local actors = {}
 
 --- GameState handlers and overrides ---
 function state:init()
-   -- table.insert(actors, 'mainActor', Actor(0, 0, 36, 36))
-   local actor = Actor(0, 0, 36, 36)
-   local actorName = 'mainActor'
+   -- table.insert(actors, 'mainObject', Object(0, 0, 36, 36))
+   local actor = Object(0, 0, 36, 36)
+   local actorName = 'mainObject'
    actors[actorName] = actor
    local player = PlayerController(actor)
    player:registerCallbacks()
@@ -33,27 +33,27 @@ function state:init()
    Signal.register("draw", function () actor:draw() end)
    Signal.register("update", function (dt) player:update(dt) end)
 
-   local actor = Actor(48, 48, 36, 36)
+   local actor = Object(48, 48, 36, 36)
    local player = PlayerController(actor)
    player:registerCallbacks()
    HC.register(actor.ps)
    Signal.register("draw", function () actor:draw() end)
    Signal.register("update", function (dt) player:update(dt) end)   
 
-   local actor = Actor(1200, 1200, 36, 36)
+   local actor = Object(1200, 1200, 36, 36)
    HC.register(actor.ps)
    Signal.register("draw", function () actor:draw() end)
    Signal.register("update", function (dt) actor:update(dt) end)   
 
    --"isometric" map image
 --   mapImage = G.newImage("assets/blue-iso.png")
-   map.init()
+   Map.init()
    cam = Camera(0, 0)
 end
 
 function state:update(dt)
    Signal.emit("update", dt)
-   cam:lookAt(actors.mainActor.ps:center())
+   cam:lookAt(actors.mainObject.ps:center())
    
 end
 
@@ -61,7 +61,7 @@ function state:draw()
    cam:attach()
 
 --   G.draw(mapImage, -1500, -1500)
-   map.drawTiles(cam:position())
+   Map.drawTiles(cam:position())
    Signal.emit("draw")
 
    cam:detach()
