--- src/keyHandler.lua
@@ -1,3 +1,0 @@
----- KeyHandler
---- Handles keypresses from the love engine and emits hump signals
-
